Enforce max size on unique table
Gives a more clear exception for two possible exceptions that occur if
the unique table becomes too large. Does allow a range of cases that
would have thrown an exception before to proceed with the largest
allowable unique table size, only failing if there is another attempt to
increase the table size.